EPMRPP-89382 || Drop support of cucumber <7 (#159)
* Drop support of cucumber <7

* feat: add tests

* feat: add tests

* fix: update jest config

* fix: use LOG_LEVELS

* fix: update findScenario function

* fix: update tests

* fix: update test name
