+++ README.md
@@ -5,7 +5,6 @@ Agent to integrate CucumberJS with ReportPortal.
 * More about [ReportPortal](http://reportportal.io/)
 
 This agent works well with cucumber versions from 7.x to 10.x.
-Documentation for legacy cucumber versions from 4.x to 6.x can be found [here](/modules/api/deprecated/README.md)
 
 ## Install agent to your project dir
