+++ modules/constants.js
@@ -14,7 +14,6 @@
  *  limitations under the License.
  */
 
-const AFTER_HOOK_URI_TO_SKIP = 'protractor-cucumber-framework';
 const STATUSES = {
   PASSED: 'passed',
   FAILED: 'failed',
@@ -84,36 +83,19 @@ const RP_EVENTS = {
   STATUS: 'rp/status',
 };
 
-// @see https://github.com/Automattic/cli-table#custom-styles
-const TABLE_CONFIG = {
-  chars: {
-    top: '',
-    'top-left': '',
-    'top-mid': '',
-    'top-right': '',
-    mid: '',
-    'left-mid': '',
-    'mid-mid': '',
-    'right-mid': '',
-    bottom: '',
-    'bottom-left': '',
-    'bottom-mid': '',
-    'bottom-right': '',
-  },
-  style: {
-    head: [],
-    border: [],
-  },
+const TEST_STEP_FINISHED_RP_MESSAGES = {
+  PENDING: "This step is marked as 'pending'",
+  UNDEFINED: 'There is no step definition found. Please verify and implement it.',
+  AMBIGUOUS: 'There are more than one step implementation. Please verify and reimplement it.',
 };
 
 module.exports = {
-  AFTER_HOOK_URI_TO_SKIP,
   RP_ENTITY_LAUNCH,
   STATUSES,
   LOG_LEVELS,
   CUCUMBER_EVENTS,
   RP_EVENTS,
-  TABLE_CONFIG,
   CUCUMBER_MESSAGES,
+  TEST_STEP_FINISHED_RP_MESSAGES,
   TEST_ITEM_TYPES,
 };
